requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
cmake==3.30.3
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dlib==19.24.6
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
gast==0.6.0
gensim==4.3.3
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
idna==3.8
inflection==0.5.1
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kneed==0.8.5
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.4
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
pyswip==0.2.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
setuptools==72.2.0
shapely==2.0.6
six==1.16.0
smart-open==7.0.4
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.66.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
